Simple Reflection System
restructure to be able to walk up the chain to find the world we might exist in.
"typeName" manually typed at the moment, would like to find an automatic way of gathering this.
SetOuter() also manually called, would like this to be built into creation methods. Only base objects should be able to have nil worlds probably?

was able to pass a main physics world reference out to terrain.

BeginPlay and OnBeginPlay concepts introduced, need to verify how these will be triggered and know when simulation is running.